requirements.txt
appnope==0.1.2
backcall==0.2.0
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
gym==0.18.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pyglet==1.5.0
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.1
six==1.15.0
torch==1.7.1
torchvision==0.8.2
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5
